% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{criar_paleta}
\alias{criar_paleta}
\title{Criar a paleta de cores}
\usage{
criar_paleta(img, n = 5)
}
\arguments{
\item{img}{caminho p/ a imagem da qual deseja criar a paleta}

\item{n}{número de cores na paleta}
}
\description{
Criar a paleta de cores
}

